+++ CHANGES.rst
@@ -1,6 +1,20 @@
 Change Log
 ==========
 
+0.3 (unreleased)
+----------------
+
+- Add ``objectify`` function on ``SQLAlchemySchemaNode`` -- use this to
+  recreate SQLAlchemy object instances from the configured mappers.
+  This new method is the opposite of ``dictify``.
+  [davidjb]
+
+
+0.2 (2013-05-16)
+----------------
+
+- No changes.
+
 0.2a1 (2012-04-09)
 ------------------